Fixed the Twitter/X link not clickable on mobile
The Twitter/X handle on the game-over screen was unclickable on mobile because a global `touchstart` listener would restart the game.

This change updates the event handler to check the tap's target. If the social link is tapped, the default browser action is allowed, and the game does not restart.